Add "Programming as Theory Building"
This was a real treat to read, and the argument from Naur was far
grander than I assumed it would be from the title. It is well-argued,
and written in 1985, and nearly 40 years later I do not believe our
profession has internalized it, but have rather spent nearly have a
century trying to ignore it while only proving Naur correct.